backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
websockets==12.0
grpcio==1.66.1
grpcio-tools==1.66.1
requests==2.32.3
numpy==2.1.1
pydantic==2.9.2
email-validator==2.2.0
tenacity==9.0.0
python-multipart==0.0.9
orjson==3.10.7
SQLAlchemy==2.0.32
asyncpg==0.29.0

# Parsing
pypdf==4.3.1
docx2txt==0.8
boto3==1.34.162
reportlab==4.2.5

# Background tasks
celery==5.3.6
redis==5.0.7

# Metrics
prometheus-client==0.20.0

# AI/ML
openai>=1.58.0
websockets>=12.0

